refactoring for readability, changed all use of uitoa_base to itoa_base

+++ power.c
@@ -6,13 +6,13 @@
 /*   By: eesaki <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/06/30 23:24:58 by eesaki            #+#    #+#             */
-/*   Updated: 2019/09/22 04:14:34 by eesaki           ###   ########.fr       */
+/*   Updated: 2019/09/28 03:56:47 by eesaki           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-long double		power(intmax_t base, intmax_t exponent)
+long double		power(int base, intmax_t exponent)
 {
 	long double	res;
